requirements.txt
asgiref==3.11.1
Django==6.0.2
django-cors-headers==4.9.0
djangorestframework==3.16.1
google-auth==2.38.0
pillow==12.1.0
psycopg2-binary==2.9.11
python-decouple==3.8
sqlparse==0.5.5
